separate out MoreOptionsTabs logic into separate helper file
